refactor: DRY - put repeating password change functionality into it's own file
